% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aapl}
\alias{aapl}
\title{Apple Share Prices}
\format{
A data frame with 124 rows and 8 variables:
\describe{
\item{symbol}{stock symbol or ticker}
\item{date}{transaction date}
\item{open}{The opening price of the day}
\item{high}{The highest price of the day}
\item{low}{The lowest price of the day}
\item{close}{The close/last price of the day}
\item{volume}{Number of shares traded}
\item{adjusted}{Adjusted close price}
}
}
\usage{
aapl
}
\description{
A dataset containing the share prices of Apple Inc. for the period 2022-01-01
to 2022-06-30.
}
\keyword{datasets}
