% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_adx.R
\name{add_adx}
\alias{add_adx}
\title{add Average Directional Index to a financial chart}
\usage{
add_adx(plot, high = "high", low = "low", n = 14, ma_type = "EMA")
}
\arguments{
\item{plot}{Needs to be a ggplot2 financial plot containing stock prices.}

\item{high}{The high column. Default is "high".}

\item{low}{The low columns. Defalut is "low".}

\item{n}{Number of periods for the ATR calculation. Default is 14.}

\item{ma_type}{Which type of moving average will be used to calculate the ATR.
The default is "EMA". Possible options are: SMA or EMA.}
}
\value{
Plots a patchwork object.
}
\description{
\code{add_adx()} will add an Average Directional Index subplot below a financial chart.
The positive direction (DIp) is plotted in green, the negative direction (DIn) is
plotted in red, and the ADX is plotted in dark blue.
}
\examples{
library(ggplot2)
library(ggttr)

data(aapl)

p <- ggplot(aapl, aes(x = date, y = adjusted)) + 
  geom_line() +
  scale_x_finance(dates = aapl$date)
  
add_adx(p, high = "high", low = "low") 
}
\seealso{
More information about the \href{https://www.investopedia.com/terms/a/adx.asp}{ADX}
indicator can be found on investopedia.
}
