% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_atr.R
\name{add_atr}
\alias{add_atr}
\title{add Average True Range indicator to a financial chart}
\usage{
add_atr(plot, high = "high", low = "low", n = 14, ma_type = "EMA")
}
\arguments{
\item{plot}{Needs to be a ggplot2 financial plot containing stock prices.}

\item{high}{The high column. Default is "high".}

\item{low}{The low columns. Defalut is "low".}

\item{n}{Number of periods for the ATR calculation. Default is 14.}

\item{ma_type}{Which type of moving average will be used to calculate the ATR.
The default is "EMA". Possible options are: SMA or EMA.}
}
\value{
Plots a patchwork object.
}
\description{
\code{add_atr()} will add an Average True Range subplot below a financial chart.
}
\examples{
library(ggplot2)
library(ggttr)

data(aapl)

p <- ggplot(aapl, aes(x = date, y = adjusted)) + 
  geom_line() +
  scale_x_finance(dates = aapl$date)
  
add_atr(p, high = "high", low = "low") 
}
\seealso{
More information about the \href{https://www.investopedia.com/terms/a/atr.asp}{ATR}
can be found on investopedia.
}
