% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cci.R
\name{add_cci}
\alias{add_cci}
\title{add Commodity Channel Index (CCI) to a financial chart}
\usage{
add_cci(plot, high = "high", low = "low", n = 20, ma_type = "SMA", c = 0.015)
}
\arguments{
\item{plot}{Needs to be a ggplot2 financial plot containing stock prices.}

\item{high}{The high column. Default is "high".}

\item{low}{The low column. Default is "low".}

\item{n}{Number of periods for the CCI calculation. Default is 20.}

\item{ma_type}{Which type of moving average will be used to calculate the CCI.
The default is "SMA". Possible options are: SMA or EMA.}

\item{c}{Constant to apply to the mean deviation. Default = 0.015.}
}
\value{
Plots a patchwork object.
}
\description{
\code{add_cci()} will add a Commodity Channel Index (CCI) subplot below a financial chart.
}
\examples{
library(ggplot2)
library(ggttr)

data(aapl)

p <- ggplot(aapl, aes(x = date, y = adjusted)) + 
  geom_line() +
  scale_x_finance(dates = aapl$date)
  
add_cci(p, high = "high", low = "low") 
}
\seealso{
More information about the \href{https://www.investopedia.com/terms/c/commoditychannelindex.asp}{CCI}
can be found on investopedia.
}
