% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_macd.R
\name{add_macd}
\alias{add_macd}
\title{Add MACD to a financial chart}
\usage{
add_macd(
  plot,
  ma_type = "EMA",
  fast = 12,
  slow = 26,
  signal = 9,
  percent = TRUE
)
}
\arguments{
\item{plot}{Needs to be a ggplot2 financial plot containing stock prices.}

\item{ma_type}{Which type of moving average will be used to calculate the macd.
The default is "EMA". Possible options are: SMA or EMA.}

\item{fast}{Number of periods for fast moving average.}

\item{slow}{Number of periods for slow moving average.}

\item{signal}{Number of periods for signal moving average.}

\item{percent}{logical; if TRUE, the percentage difference between the fast
and slow moving averages is returned, otherwise the difference between the
respective averages is returned.}
}
\value{
Plots a patchwork object.
}
\description{
\code{add_macd()} will add a Moving Average Convergence Divergence indicator sub plot below a financial chart.
}
\examples{
library(ggplot2)
library(ggttr)

data(aapl)

p <- ggplot(aapl, aes(x = date, y = adjusted)) + 
  geom_line() +
  scale_x_finance(dates = aapl$date)
  
add_macd(p) 
}
\seealso{
More information about \href{https://www.investopedia.com/terms/m/macd.asp}{MACD}
can be found on investopedia.
}
