% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_momentum_roc.R
\name{add_momentum}
\alias{add_momentum}
\title{Add a Momentum indicator to a financial chart}
\usage{
add_momentum(plot, n = 1)
}
\arguments{
\item{plot}{Needs to be a ggplot2 financial plot containing stock prices.}

\item{n}{Number of periods to use. Default is 1.}
}
\value{
Plots a patchwork object.
}
\description{
The momentum indicator plots the difference between 2 observations.
}
\examples{
library(ggplot2)
library(ggttr)

data(aapl)

p <- ggplot(aapl, aes(x = date, y = adjusted)) + 
  geom_line() +
  scale_x_finance(dates = aapl$date)
  
add_momentum(p) 
}
\seealso{
More information about \href{https://www.investopedia.com/terms/m/momentum.asp}{momentum}
can be found on investopedia.
}
