% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_momentum_roc.R
\name{add_roc}
\alias{add_roc}
\title{Add a ROC indicator plot to a financial chart}
\usage{
add_roc(plot, n = 1, type = c("continuous"))
}
\arguments{
\item{plot}{Needs to be a ggplot2 financial plot containing stock prices.}

\item{n}{Number of periods to use. Default is 1.}

\item{type}{Compounding type; either "continuous" (the default) or "discrete".}
}
\value{
Plots a patchwork object.
}
\description{
The Rate of Change indicator plots the percentage difference between 2 observations.
}
\examples{
library(ggplot2)
library(ggttr)

data(aapl)

p <- ggplot(aapl, aes(x = date, y = adjusted)) + 
  geom_line() +
  scale_x_finance(dates = aapl$date)
  
add_roc(p) 
}
\seealso{
More information about \href{https://www.investopedia.com/terms/p/pricerateofchange.asp}{ROC}
can be found on investopedia.
}
