% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rsi.R
\name{add_rsi}
\alias{add_rsi}
\title{Add RSI to a financial chart}
\usage{
add_rsi(plot, n = 14)
}
\arguments{
\item{plot}{Needs to be a ggplot2 financial plot containing stock prices.}

\item{n}{Number of periods for the RSI calculation. Default is 14.}
}
\value{
Plots a patchwork object.
}
\description{
Add RSI to a financial chart
}
\examples{
library(ggplot2)
library(ggttr)

data(aapl)

p <- ggplot(aapl, aes(x = date, y = adjusted)) + 
  geom_line() +
  scale_x_finance(dates = aapl$date)
  
add_rsi(p) 
}
