% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stoch.R
\name{add_stoch}
\alias{add_stoch}
\title{add a Stochastic Oscillator to a financial chart}
\usage{
add_stoch(
  plot,
  high = "high",
  low = "low",
  fastK = 14,
  fastD = 3,
  slowD = 1,
  ma_type = "EMA"
)
}
\arguments{
\item{plot}{Needs to be a ggplot2 financial plot containing stock prices.}

\item{high}{The high column. Default is "high".}

\item{low}{The low column. Default is "low".}

\item{fastK}{Number of periods for fast \%K. Default is 14.}

\item{fastD}{Period of the fast average. Default is 3.}

\item{slowD}{period of the slow average. Default is 1.}

\item{ma_type}{Which type of moving average will be used to calculate the Stochastic Oscillator.
The default is "EMA". Possible options are: SMA or EMA.}
}
\value{
Plots a patchwork object.
}
\description{
\code{add_stoch()} will add a Stochastic Oscillator subplot below a financial chart.
The default settings are the same as in Interactive Brokers.
}
\examples{
library(ggplot2)
library(ggttr)

data(aapl)

p <- ggplot(aapl, aes(x = date, y = adjusted)) + 
  geom_line() +
  scale_x_finance(dates = aapl$date)
  
add_stoch(p, high = "high", low = "low") 
}
\seealso{
More information about the \href{https://www.investopedia.com/terms/s/stochasticoscillator.asp}{Stochastic Oscillator}
can be found on investopedia.
}
