% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_donchian.R
\name{geom_donchian}
\alias{geom_donchian}
\title{Plot a Donchian Channel}
\usage{
geom_donchian(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  n = 20,
  include_lag = FALSE,
  colour_ma = "blue",
  colour_bands = "red",
  alpha = 0.2,
  fill = "grey",
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{n}{Number of periods to average over. Defaults to 20.}

\item{include_lag}{Should values be lagged so that today's prices are not included in the calculation?}

\item{colour_ma, colour_bands}{Line colours of the Bollinger Bands and the moving average line.}

\item{alpha}{Transparancy for the Bollinger Band ribbon.}

\item{fill}{Fill colour between the outer bands. Default is grey.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
The \code{geom_donchian()} function will plot the Donchian channel around a security's price.
}
\examples{
library(ggplot2)
library(ggttr)

data(aapl)

ggplot(aapl, aes(x = date, y = adjusted)) + 
  geom_line() +
  geom_donchian(aes(high = high, low = low))
  
}
\seealso{
More information about \href{https://www.investopedia.com/terms/d/donchianchannels.asp}{Donchian channels}
can be found on investopedia.
}
