% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ma.R
\name{geom_ma}
\alias{geom_ma}
\title{Plot Moving Averages}
\usage{
geom_ma(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  ma_type = "SMA",
  n = 10,
  wilder = FALSE,
  ratio = NULL,
  v = 1,
  wts = 1:n,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{ma_type}{Which type of moving average will be used to calculate the moving
average. The default is "SMA". Possible options are: SMA, EMA, DEMA, WMA, ZLEMA
and HMA. See \code{\link[TTR:MovingAverages]{TTR::SMA()}} for more information.}

\item{n}{Number of periods to average over. Must be between 1 and \code{nrow(x)}, inclusive.
Defaults to 10.}

\item{wilder}{logical; if \code{TRUE}, a Welles Wilder type EMA will be calculated.
Default value is \code{FALSE}.}

\item{ratio}{A smoothing/decay ratio. Default is \code{NULL}.}

\item{v}{The 'volume factor' (a number between \verb{[0,1]}). Default is 1.}

\item{wts}{Vector of weights. Length of wts vector must equal the length of x, or n (the default).}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
The \code{geom_ma()} function plots a moving average on a financial chart. You can choose from several
moving averages. The details of each moving average can be found in the TTR package.
}
\examples{
library(ggplot2)
library(ggttr)

data(aapl)

ggplot(aapl, aes(x = date, y = adjusted)) + 
  geom_line() +
  geom_ma(ma_type = "EMA")

}
