% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggttr-package.R
\docType{package}
\name{ggttr-package}
\alias{ggttr}
\alias{ggttr-package}
\title{ggttr: Technical Trading Rules for ggplot2}
\description{
The purpose of ggttr is to add technical trading rules to ggplot2. This adds the functionality of plotting technical trading rules, available in quantmod via the `chartSeries()` and `addTA()` functions, to ggplot2.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
