% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_finance.R
\name{scale_x_finance}
\alias{scale_x_finance}
\title{Scale for removing non-business days from ggplot}
\usage{
scale_x_finance(dates, ...)
}
\arguments{
\item{dates}{A column containing the dates}

\item{...}{Other arguments passed on to scale_x_continous()}
}
\description{
This scale will ignore weekends and holidays that are present in stock data.
}
\details{
ggplot tends to show missing dates in the plots when plotting data like daily
stock data. It adds weekends and other missing days that (most) financial software
will ignore as this is the standard way of plotting financial data.
}
\examples{
library(ggplot2)
library(ggttr)

data(aapl)
ggplot(aapl, aes(x = date, y = adjusted)) + 
  geom_line() + 
  scale_x_finance(dates = aapl$date)
}
